+++ docs/release_notes.rst
@@ -1,6 +1,10 @@
 Release Notes
 -------------
 
+- 1.1.36 `July 18, 2022`
+    * Updated support for Google OAuth 2.0 and App Engine
+    * Ensure `anndata`_ layers are stored in sparse csc format in `cirro prepare_data`
+
 - 1.1.35 `July 1, 2022`
     * Fixed error when selecting more than one layer in `anndata`_ layers